# apps/backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/lessons": {
      "post": {
        "operationId": "LessonsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lessons"
        ]
      },
      "get": {
        "operationId": "LessonsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lessons"
        ]
      }
    },
    "/lessons/{id}": {
      "get": {
        "operationId": "LessonsController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lessons"
        ]
      },
      "patch": {
        "operationId": "LessonsController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lessons"
        ]
      },
      "delete": {
        "operationId": "LessonsController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lessons"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-Tenant ABAI API",
...
    "version": "1.0.0",
    "contact": {
      "name": "ABAI Support Team",
      "url": "https://abai.edu.kz",
      "email": "support@abai.edu.kz"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.abai.edu.kz",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Введите JWT токен",
        "in": "header"
      }
    },
...
  }
}